% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_performance.R
\name{boxplot_performance}
\alias{boxplot_performance}
\title{Box Plot Performance of a set of configurations}
\usage{
boxplot_performance(
  experiments,
  allElites = NULL,
  type = "all",
  first_is_best = FALSE,
  rpd = TRUE,
  show_points = TRUE,
  best_color = "#08bfaa",
  x_lab = "Configurations",
  filename = NULL
)
}
\arguments{
\item{experiments}{Experiment matrix obtained from irace training or testing data. Configurations 
in columns and instances in rows. As in irace, column names (configurations ids) 
should be characters.}

\item{allElites}{List or vector of configuration ids, (default NULL). These configurations
should be included in the plot. If the argument is not provided all configurations
in experiments are included. If allElites is a vector all configurations are
assumed without iteration unless argument type="ibest" is provided, in which case
each configuration is assumed to be from a different iteration. If allElites 
is a list, each element of the  list is assumed as an iteration.}

\item{type}{String, (default "all") possible values are "all", "ibest" or "best". "all" 
shows all the selected configurations showing iterations if the information 
is provided. "best" shows the elite configurations of the last iteration, note
that the best configuration is always assumed to be first in the vector.
"ibest" shows the elite configurations of each iteration, note
that the best configuration is always assumed to be first in the vector of 
each iteration.}

\item{first_is_best}{Boolean (default TRUE) Enables the display in a different color the best configuration
identified as the first one in a vector. If FALSE, all configurations are shown
in the same color.}

\item{rpd}{(\code{logical(1)}) TRUE to plot performance as the relative
percentage deviation to best results per instance, FALSE to plot raw
performance.}

\item{show_points}{Logical, (default TRUE) TRUE to plot performance points together with the box plot.}

\item{best_color}{String, (default ""#08bfaa"") color to display best configurations.}

\item{x_lab}{String, (default "Configurations") label for the x axis.}

\item{filename}{(\code{character(1)}) File name to save the plot, for example \code{"~/patch/example/filename.png"}.}
}
\value{
box plot
}
\description{
The `boxplot_performance` function creates a box plot that displays the 
performance of a set of configurations which can be displayed by iteration. 

The performance data is obtained from the experiment matrix provided in the 
experiments argument. The configurations can be selected using the allElites
argument and this argument can be also used to define the iteration of each
elite configuration was evaluated.
}
\examples{
boxplot_performance(iraceResults$experiments, iraceResults$allElites)
\dontrun{ 
boxplot_performance(iraceResults$testing$experiments, iraceResults$iterationElites)
}
}
