% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_training.R
\name{boxplot_training}
\alias{boxplot_training}
\title{Box Plot}
\usage{
boxplot_training(
  iraceResults,
  numberIteration = NULL,
  idConfigurations = NULL,
  rpd = TRUE,
  fileName = NULL
)
}
\arguments{
\item{iraceResults}{The data generated when loading the Rdata file created by irace}

\item{numberIteration}{Numeric, It is the number referring to the iteration that you want to graph,
the range of these varies according to the Rdata used (example: numberIteration = 5)}

\item{idConfigurations}{Numeric vector, you need to put the configurations you want to analyze
(example: idConfigurations = c(20,50,100,300,500,600,700))}

\item{rpd}{Logical (default TRUE) to fit through an equation of minimum percentage distance
between the values of each row of all configurations}

\item{fileName}{String, A pdf will be created in the location and with the
assigned name (example: "~/patch/example/filename")}
}
\value{
box plot
}
\description{
Create a graphic of box plot of irace using the best quality id
By default the graph of the last iteration is shown
}
\examples{
boxplot_training(iraceResults)
boxplot_training(iraceResults,numberIteration = 5)
boxplot_training(iraceResults,idConfigurations = c(20,50,100,300,500,600,700))

}
