% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_config.R
\name{distance_config}
\alias{distance_config}
\title{Distance between configurations}
\usage{
distance_config(irace_results, id_configurations, t = 0.05)
}
\arguments{
\item{irace_results}{The data generated when loading the \code{Rdata} file created by \code{irace}}

\item{id_configurations}{Numeric vector, you need to put the settings you
want to compare, only 2 values are allowed (example: idVector = c(806,809))}

\item{t}{Numeric, It is a percentage factor that will determine the range of difference
between settings (example: t = 0.05 is equivalent to 5 percent)}
}
\value{
numeric
}
\description{
Calculate the difference between two settings. Calculate the difference between two configurations. The greater
the number, the greater the difference
}
\examples{
NULL
}
