% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_iteration.R
\name{distance_iteration}
\alias{distance_iteration}
\title{Distance Iteration Plot}
\usage{
distance_iteration(iraceResults, type = "both", t = 0.05, fileName = NULL)
}
\arguments{
\item{iraceResults}{The data generated when loading the Rdata file created by irace}

\item{type}{String, either "line", "boxplot" or "both". by default it is "both" will show both graphics, "line" which will show a plot of
points and lines, "boxplot" will show a box plot}

\item{t}{Numeric, It is a percentage factor that will determine the range of difference
between settings (example: t = 0.05 is equivalent to 5 percent)}

\item{fileName}{String, A pdf will be created in the location and with the assigned
name (example: "~/patch/example/filename")}
}
\value{
line or box plot
}
\description{
Shows the mean of the difference between the configurations that were run for each iteration
}
\examples{
NULL
}
