% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parameters_names.R
\name{get_parameters_names}
\alias{get_parameters_names}
\title{Parameters Names}
\usage{
get_parameters_names(iraceResults)
}
\arguments{
\item{iraceResults}{The data generated when loading the Rdata file created by irace}
}
\value{
String vector
}
\description{
The function returns a string vector with the names of
the document parameters iraceResults
}
\examples{
get_parameters_names(iraceResults)
}
