% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model.R
\name{getCategoricalModel}
\alias{getCategoricalModel}
\title{Categorical model generation}
\usage{
getCategoricalModel(irace_results, param_name)
}
\arguments{
\item{irace_results}{The data generated when loading the \code{Rdata} file created by \code{irace}}

\item{param_name}{String, parameter to be included in the plot (example: param_name = "algorithm"))}
}
\value{
data frame with columns "iteration", "elite", "parameter", "value", "prob"
}
\description{
The `getCategoricalModel` function rebuilds the probabilities of
the sampling models used in irace to generate configurations in each
iteration.
}
\examples{
NULL
}
