% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model.R
\name{getNumericalModel}
\alias{getNumericalModel}
\title{Numerical model generation}
\usage{
getNumericalModel(irace_results, param_name)
}
\arguments{
\item{irace_results}{The data generated when loading the \code{Rdata} file created by \code{irace}}

\item{param_name}{String, parameter to be included in the plot (example: param_name = "algorithm"))}
}
\value{
data frame with columns "iteration", "elite", "parameter", "mean", "sd"
}
\description{
The `getNumericalModel` function rebuilds the sampling distribution parameters
of the models used by irace to sampling configurations during the configuration 
process.
}
\examples{
NULL
}
