% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{has_testing_data}
\alias{has_testing_data}
\title{Check if the results object generated by irace has data about the testing phase.}
\usage{
has_testing_data(irace_results)
}
\arguments{
\item{irace_results}{The data generated when loading the \code{.Rdata} file created by \code{irace} (or the filename of that file).}
}
\value{
\code{logical(1)}
}
\description{
Check if the results object generated by irace has data about the testing phase.
}
