% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{heatmap}
\alias{heatmap}
\title{Heat Map Plot}
\usage{
heatmap(irace_results, file_name = NULL)
}
\arguments{
\item{irace_results}{The data generated when loading the Rdata file created by irace}

\item{file_name}{string, A pdf will be created in the location and with the
assigned name (example: "~/patch/example/file_name")}
}
\value{
heatmap plot
}
\description{
The function will return a heat map plot of all experimental data configurations
}
\examples{
heatmap(iraceResults)
}
