% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iboxplot_test.R
\name{iboxplot_test}
\alias{iboxplot_test}
\title{Box Plot Testing}
\usage{
iboxplot_test(iraceResults, type = "all", rpd = TRUE, fileName = NULL)
}
\arguments{
\item{iraceResults}{The data generated when loading the Rdata file created by irace}

\item{type}{String, either "all", "ibest" or "best". By default it is "all" which shows all the configurations,
"best" shows the best configurations of each iteration and
"last" shows the configurations of the last iteration}

\item{rpd}{Logical (default TRUE) to fit through an equation of minimum percentage distance between
the values of each row of all configurations}

\item{fileName}{String, A pdf will be created in the location and with the assigned
name (example: "~/patch/example/filename")}
}
\value{
box plot
}
\description{
The function will return a box plot, using the data generated in the test
settings coloring the best configuration in each iteration
}
\examples{
NULL
}
