% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibp.R
\name{ibp}
\alias{ibp}
\title{Graphic box plot}
\usage{
ibp(iraceResults)
}
\arguments{
\item{iraceResults}{The data generated when loading the Rdata file created by irace}
}
\value{
plot
}
\description{
Create a graphic of box plot from the latest iteration
of irace using the best quality id
}
\examples{
NULL
}
