% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iconf_instances.R
\name{iconf_instances}
\alias{iconf_instances}
\title{The configurations by iteration and instance}
\usage{
iconf_instances(iraceResults, rpd = FALSE, fileName = NULL)
}
\arguments{
\item{iraceResults}{The data generated when loading the Rdata file created by irace}

\item{rpd}{Logical (default FALSE) to fit through an equation of minimum percentage distance
between the values of each row of all configurations}

\item{fileName}{String, A pdf will be created in the location and with the
assigned name (example: "~/patch/example/filename")}
}
\value{
plot
}
\description{
A graph is created with all the settings and instance of the training data
}
\examples{
NULL
}
