% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idistance_iteration.R
\name{idistance_iteration}
\alias{idistance_iteration}
\title{Distance Iteration Plot}
\usage{
idistance_iteration(iraceResults, type = "line", t = 0.05, fileName = NULL)
}
\arguments{
\item{iraceResults}{The data generated when loading the Rdata file created by irace}

\item{type}{String, either "line" or "boxplot". by default it is "line" which will show a plot of
points and lines, "boxplot" will show a box plot}

\item{t}{Numeric, It is a percentage factor that will determine the range of difference
between settings (example: t = 0.05 is equivalent to 5 percent)}

\item{fileName}{String, A pdf will be created in the location and with the assigned
name (example: "~/patch/example/filename")}
}
\value{
line or box plot
}
\description{
Shows the mean of the difference between the configurations that were run for each iteration
}
\examples{
NULL
}
