% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iheatmap.R
\name{iheatmap}
\alias{iheatmap}
\title{Heat Map Plot}
\usage{
iheatmap(iraceResults, fileName = NULL)
}
\arguments{
\item{iraceResults}{The data generated when loading the Rdata file created by irace}

\item{fileName}{string, A pdf will be created in the location and with the
assigned name (example: "~/patch/example/filename")}
}
\value{
heatmap plot
}
\description{
The function will return a heat map plot of all experimental data configurations
}
\examples{
NULL
}
