% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iparallel_coord.R
\name{iparallel_coord}
\alias{iparallel_coord}
\title{Parallel Coordinates Plot interactive}
\usage{
iparallel_coord(irace_results, id_configuration = NULL, param_names = NULL)
}
\arguments{
\item{irace_results}{The data generated when loading the Rdata file created by irace}

\item{id_configuration}{Numeric vector, you need to put the configurations you want to analyze
(example: id_configuration = c(20,50,100,300,500,600,700))}

\item{param_names}{String vector, you need to put the parameters you want to analyze
(example: param_names = c("algorithm","alpha","rho","q0","rasrank"))}
}
\value{
parallel cordinates plot
}
\description{
The function will returns a parallel coordinates plot
interactive in shinyApp allowing the analysis of the set of parameters
allowing the visualization of the data and filter by iteration
}
\examples{
NULL
}
