% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iparameter_compare_iteration.R
\name{iparameter_compare_iteration}
\alias{iparameter_compare_iteration}
\title{Frequency and Density plot based on its iteration}
\usage{
iparameter_compare_iteration(iraceResults, parameter, fileName = NULL)
}
\arguments{
\item{iraceResults}{The data generated when loading the Rdata file created by irace}

\item{parameter}{String, value of the parameter to be analyzed (example: parameter = "alpha")}

\item{fileName}{String, A pdf will be created in the location and with the
assigned name (example: "~/patch/example/filename")}
}
\value{
frequency and density plot
}
\description{
The function will return two plots, one of frequency and
one of density of the numerical parameter, in case of being a categorical parameter
it will only show one of frequency
}
\examples{
NULL
}
