% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iparameter_freq.R
\name{iparameter_freq}
\alias{iparameter_freq}
\title{Parameter Frequency and density Plot}
\usage{
iparameter_freq(iraceResults, param_names = NULL, fileName = NULL)
}
\arguments{
\item{iraceResults}{The data generated when loading the Rdata file created by irace}

\item{param_names}{It's of type vector
Is a vector with the parameters to be displayed in the plot}

\item{fileName}{It's of type string
You must place the path where you want to save the file and its name without the extension pdf}
}
\value{
Frequency and/or density plot
}
\description{
The function will return a frequency and density plot,
for categorical parameters (string) one of frequency is created,
in case of numerical parameters it will show a histogram and its density
}
\examples{
NULL
}
