% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iparcoord.R
\name{iparcoord}
\alias{iparcoord}
\title{Parallel Coordinates Plot}
\usage{
iparcoord(iraceResults, idIteration = NULL, fileName = NULL)
}
\arguments{
\item{iraceResults}{The data generated when loading the Rdata file created by irace}

\item{idIteration}{It is a vector with id values that you want to graph}

\item{fileName}{It's of type string
You must place the path where you want to save the file and its name without the extension pdf}
}
\value{
plot
}
\description{
Parallel Coordinates Plot
}
\examples{
NULL
}
