% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irace_summarise.R
\name{irace_summarise}
\alias{irace_summarise}
\title{Summarise the results of a run of irace}
\usage{
irace_summarise(irace_results)
}
\arguments{
\item{irace_results}{The data generated when loading the \code{.Rdata} file created by \code{irace} (or the filename of that file).}
}
\value{
\code{list()}
}
\description{
Summarise the results of a run of irace
}
\examples{
irace_results <- read_logfile(system.file(package="iraceplot", "exdata",
                                         "guide-example.Rdata", mustWork = TRUE))
irace_summarise(irace_results)

}
\author{
Manuel López-Ibáñez
}
