% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iraceplot-package.R
\docType{package}
\name{iraceplot-package}
\alias{iraceplot-package}
\title{The iraceplot package: \packageTitle{iraceplot}}
\description{
\packageDescription{iraceplot}
}
\details{
License: MIT + file LICENSE
}
\examples{
###########################################
# These examples illustrate how the various
# functions of the package works.
###########################################

# First we need to locate the data file generated by irace,
# which will be used for the various graphics




}
\author{
Maintainers: Pablo Oñate Marín and Leslie Pérez Cáceres
        \email{pablo.onate.m@gmail.com}
}
\keyword{automatic}
\keyword{configuration}
\keyword{package}
\keyword{plot}
