% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iraceResults}
\alias{iraceResults}
\title{Data created by the irace package
Large list with 11 elements}
\format{
An object of class \code{list} of length 11.
}
\usage{
iraceResults
}
\description{
Data created by the irace package
Large list with 11 elements
}
\keyword{datasets}
