% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iscatter_test.R
\name{iscatter_test}
\alias{iscatter_test}
\title{Scatter Plot Testing}
\usage{
iscatter_test(iraceResults, idVector, rpd = FALSE, fileName = NULL)
}
\arguments{
\item{iraceResults}{The data generated when loading the Rdata file created by irace}

\item{idVector}{String vector, you need to put the elite settings you
want to compare, only 2 values are allowed (example: idVector = c("92","119"))}

\item{rpd}{logical(default FALSE) to fit through an equation of minimum percentage distance
between the values of each row of all configurations}

\item{fileName}{String, A pdf will be created in the location and with the assigned
name (example: "~/patch/example/filename")}
}
\value{
scatter plot
}
\description{
The function will return a scatter plot
comparing two elite configurations in testing
}
\examples{
NULL
}
