% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_cat.R
\name{parallel_cat}
\alias{parallel_cat}
\title{Parallel Cordinate Category}
\usage{
parallel_cat(
  irace_results,
  id_configuration = NULL,
  param_names = NULL,
  iterations = NULL,
  pdf_all_parameters = FALSE,
  file_name = NULL
)
}
\arguments{
\item{irace_results}{The data generated when loading the \code{Rdata} file created by \code{irace}}

\item{id_configuration}{Numeric vector, you need to put the configurations you want to analyze
(example: id_configuration = c(20,50,100,300,500,600,700))}

\item{param_names}{String vector, you need to put the parameters you want to analyze
(example: param_names = c("algorithm","alpha","rho","q0","rasrank"))}

\item{iterations}{NUmeric vector, you need to put the iterations you want to analyze
(example: iterations = c(1,4,5))}

\item{pdf_all_parameters}{logical (default FALSE), If I want to create a pdf with all the parameters,
I must put TRUE, otherwise it will be created only with the default
parameters (15 or less) or those entered.}

\item{file_name}{A pdf will be created in the location and with
the assigned name (example: "~/patch/example/file_name")}
}
\value{
parallel coordinate category plot
}
\description{
The parallel_coord_category function will return a graph of categorical
parallel coordinates allowing the analysis of the set of parameters
allowing the visualization of the data and the filtering by iteration
}
\examples{
parallel_cat(iraceResults)
parallel_cat(iraceResults,id_configuration = c(20,50,100,300,500,600,700))
parallel_cat(iraceResults, param_names = c("algorithm","alpha","rho","q0","rasrank"))
parallel_cat(iraceResults, iterations =  c(1,4,6))
}
