% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_coord.R
\name{parallel_coord}
\alias{parallel_coord}
\title{Parallel Coordinates Plot static}
\usage{
parallel_coord(
  irace_results,
  id_configuration = NULL,
  param_names = NULL,
  iterations = NULL,
  only_elite = TRUE,
  plot_all_parameters = FALSE,
  file_name = NULL
)
}
\arguments{
\item{irace_results}{The data generated when loading the \code{Rdata} file created by \code{irace}}

\item{id_configuration}{Numeric vector, configurations ids whose performance should be included in the plot
(example: id_configuration = c(20,50,100,300,500,600,700))}

\item{param_names}{String vector, names of the parameters that should be included in the plot
(example: param_names = c("algorithm","alpha","rho","q0","rasrank"))}

\item{iterations}{Numeric vector, iteration number that should be included in the plot
(example: iterations = c(1,4,5))}

\item{only_elite}{logical (default TRUE), only print elite configurations (argument ignored when id_configuration is provided)}

\item{plot_all_parameters}{logical (default FALSE), TRUE if all parameter should be included or 
FALSE if the first set of 15 parameters provided or in the data file should
be displayed.}

\item{file_name}{String, file name to save plot (example: "~/patch/example/file_name.png")}
}
\value{
parallel coordinates plot
}
\description{
The parallel_coord function returns a parallel cordinates plot
of a set of configurations
}
\examples{
parallel_coord(iraceResults)
parallel_coord(iraceResults, id_configuration = c(20, 50, 100, 300, 500, 600, 700))
parallel_coord(iraceResults, param_names = c("algorithm", "alpha", "rho", "q0", "rasrank"))
parallel_coord(iraceResults, iterations = c(1, 4, 6))
}
