% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_frequency.R
\name{parameter_frequency}
\alias{parameter_frequency}
\title{Parameter Frequency and density Plot}
\usage{
parameter_frequency(iraceResults, param_names = NULL, fileName = NULL)
}
\arguments{
\item{iraceResults}{The data generated when loading the Rdata file created by irace}

\item{param_names}{String vector, A set of parameters to be plotted
(example: param_names = c("algorithm","alpha","rho","q0","rasrank"))}

\item{fileName}{String, A pdf will be created in the location and with the assigned
name (example: "~/patch/example/filename")}
}
\value{
Frequency and/or density plot
}
\description{
The function will return a frequency and density plot,
for categorical parameters (string) one of frequency is created,
in case of numerical parameters it will show a histogram and its density
}
\examples{
NULL
}
