% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_frequency_iteration.R
\name{parameter_frequency_iteration}
\alias{parameter_frequency_iteration}
\title{Frequency and Density plot based on its iteration}
\usage{
parameter_frequency_iteration(iraceResults, parameter, fileName = NULL)
}
\arguments{
\item{iraceResults}{The data generated when loading the Rdata file created by irace}

\item{parameter}{String, value of the parameter to be analyzed (example: parameter = "algorithm")}

\item{fileName}{String, A pdf will be created in the location and with the assigned
name (example: "~/patch/example/filename")}
}
\value{
bar plot
}
\description{
The function will return a frequency plot used
for categorical data (its values are string, show a bar plot) or
numeric data (show a histogram and density plot) by each iteration
}
\examples{
NULL
}
