% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model.R
\name{plotCategoricalModel}
\alias{plotCategoricalModel}
\title{Plot a categorical model}
\usage{
plotCategoricalModel(model_data, domain)
}
\arguments{
\item{model_data}{String, data frame obtained from the `getCategoricalModel` function}

\item{domain}{String Vector, domain of the parameter whose model will be plotted}
}
\value{
bar plot
}
\description{
The `plotCategoricalModel` function creates a stacked bar plot showing
the sampling probabilities of the parameter values for the elite
configurations in the iterations of the configuration process.
}
