% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model.R
\name{plotNumericalModel}
\alias{plotNumericalModel}
\title{Plot a categorical model}
\usage{
plotNumericalModel(iteration, model_data, domain, xlabel_iteration)
}
\arguments{
\item{iteration}{Numeric, iteration that should considered in the plot}

\item{model_data}{String, data frame obtained from the `getNumericalModel` function}

\item{domain}{Numeric vector, domain of the parameter whose model will be plotted}

\item{xlabel_iteration}{Numeric, iteration in which the x axis labels should be included}
}
\value{
sampling distribution plot
}
\description{
The `plotNumericalModel` function creates a sampling distributions plot of the
numerical parameters for the elite configurations of an iteration.

This plot shows de density function of the truncated normal distributions
associated to each parameter for each elite configuration.
}
