% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{read_logfile}
\alias{has_testing_data}
\alias{read_ablogfile}
\alias{irace_summarise}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{irace}{\code{\link[irace]{has_testing_data}}, \code{\link[irace]{irace_summarise}}, \code{\link[irace]{read_ablogfile}}, \code{\link[irace]{read_logfile}}}
}}

