% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{Report Iraceplot}
\usage{
report(irace_results, filename)
}
\arguments{
\item{irace_results}{The data generated when loading the \code{Rdata} file created by \code{irace}}

\item{filename}{(\code{character(1)})
String, file name indicating where to save the report (example: "~/path-to/filename")}
}
\value{
document
}
\description{
The \code{report} function creates an HTML report of the most relevant irace log
data.
}
\details{
This report provides general statistics and plots that show the best configurations
and their performance.
}
\examples{
NULL
}
