% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_density.R
\name{sampling_density}
\alias{sampling_density}
\title{Density plot based on its iteration}
\usage{
sampling_density(iraceResults, parameter, fileName = NULL)
}
\arguments{
\item{iraceResults}{The data generated when loading the Rdata file created by irace}

\item{parameter}{String, value of the parameter to be analyzed (example: parameter = "alpha")}

\item{fileName}{String, A pdf will be created in the location and with the
assigned name (example: "~/patch/example/filename")}
}
\value{
density plot
}
\description{
The function will return a density plot for parameters numeric data
in case of placing a categorical parameter, it will show a bar plot
}
\examples{
NULL
}
