% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_distance.R
\name{sampling_distance}
\alias{sampling_distance}
\title{Distance Iteration Plot}
\usage{
sampling_distance(irace_results, type = "both", t = 0.05, file_name = NULL)
}
\arguments{
\item{irace_results}{The data generated when loading the \code{Rdata} file created by \code{irace}}

\item{type}{String, either "line", "boxplot" or "both". by default it is "both" will show both graphics, "line" which will show a plot of
points and lines, "boxplot" will show a box plot}

\item{t}{Numeric, It is a percentage factor that will determine the range of difference
between settings (example: t = 0.05 is equivalent to 5 percent)}

\item{file_name}{String, A pdf will be created in the location and with the assigned
name (example: "~/patch/example/file_name")}
}
\value{
line or box plot
}
\description{
Shows the mean of the difference between the configurations that were run for each iteration
}
\examples{
# sampling_distance(iraceResults)
# sampling_distance(iraceResults, type = "boxplot", t=0.2)
}
