% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_frequency.R
\name{sampling_frequency}
\alias{sampling_frequency}
\title{Parameter Frequency and density Plot}
\usage{
sampling_frequency(
  irace_results,
  param_names = NULL,
  n = NULL,
  file_name = NULL
)
}
\arguments{
\item{irace_results}{The data generated when loading the \code{Rdata} file created by \code{irace}}

\item{param_names}{String vector, A set of parameters to be included
(example: param_names = c("algorithm","alpha","rho","q0","rasrank"))}

\item{n}{Numeric, for scenarios with large parameter sets, it selected a subset 
of 9 parameters. For example, n=1 selects the first 9 (1 to 9) parameters, n=2 selects
the next 9 (10 to 18) parameters and so on.}

\item{file_name}{String,  file name to save plot. If there are more than 9 parameters, 
a pdf file extension is recommended as it allows to create a multi-page
document. Otherwise, you can use the n argument of the function to generate 
the plot of a subset of the parameters. (example: "~/path/to/file_name.pdf")}
}
\value{
Frequency and/or density plot
}
\description{
The function will return a frequency or density plot,
for categorical parameters a frequency plot is created,
in case of numerical parameters a histogram and density 
plot is created.
}
\examples{
sampling_frequency(iraceResults)
sampling_frequency(iraceResults, n = 2)
sampling_frequency(iraceResults, param_names = c("algorithm", "alpha", "rho", "q0", "rasrank"))
}
