% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_frequency.R
\name{sampling_frequency2}
\alias{sampling_frequency2}
\title{Parameter Frequency and density Plot}
\usage{
sampling_frequency2(
  configurations,
  parameters,
  param_names = NULL,
  n = NULL,
  filename = NULL
)
}
\arguments{
\item{configurations}{Data frame, configurations in \code{irace} format
(example: \code{configurations = iraceResults$allConfigurations})}

\item{parameters}{List, parameter object in irace format
(example: \code{configurations = iraceResults$parameters})}

\item{param_names}{String vector, A set of parameters to be included
(example: \code{param_names = c("algorithm","alpha","rho","q0","rasrank")})}

\item{n}{Numeric, for scenarios with large parameter sets, it selects a subset
of 9 parameters. For example, \code{n=1} selects the first 9 (1 to 9) parameters, n=2 selects
the next 9 (10 to 18) parameters and so on.}

\item{filename}{(\code{character(1)}) File name to save the plot, for example \code{"~/patch/example/filename.png"}.}
}
\value{
Frequency and/or density plot
}
\description{
The \code{sampling_frequency2} function creates a frequency or density plot that
depicts the sampling represented by the set of configurations provided.

For categorical parameters a frequency plot is created, while for numerical
parameters a histogram and density plots are created. The plots are shown in
groups of maximum 9, the parameters included in the plot can be specified by
setting the param_names argument.
}
\note{
If there are more than 9 parameters, a pdf file extension is
recommended as it allows to create a multi-page document. Otherwise, you
can use the \code{n} argument of the function to generate the plot of a subset
of the parameters.
}
\examples{
sampling_frequency2(iraceResults$allConfigurations, iraceResults$parameters)
sampling_frequency2(iraceResults$allConfigurations, iraceResults$parameters, n = 2)
sampling_frequency2(iraceResults$allConfigurations, iraceResults$parameters, 
                    param_names = c("alpha"))
sampling_frequency2(iraceResults$allConfigurations, iraceResults$parameters, 
                    param_names = c("algorithm", "alpha", "rho", "q0", "rasrank"))
}
