% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_model_compare.R
\name{sampling_model_compare}
\alias{sampling_model_compare}
\title{Frequency and Density plot based on its iteration}
\usage{
sampling_model_compare(irace_results, parameter, file_name = NULL)
}
\arguments{
\item{irace_results}{The data generated when loading the Rdata file created by irace}

\item{parameter}{String, value of the parameter to be analyzed (example: parameter = "alpha")}

\item{file_name}{String, A pdf will be created in the location and with the
assigned name (example: "~/patch/example/file_name")}
}
\value{
frequency and density plot
}
\description{
The function will return two plots, one of frequency and
one of density of the numerical parameter, in case of being a categorical parameter
it will only show one of frequency
}
\examples{
sampling_model_compare(iraceResults, parameter = "alpha")
}
