% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_pie.R
\name{sampling_pie}
\alias{sampling_pie}
\title{Sunburst Plot}
\usage{
sampling_pie(irace_results, param_names = NULL, file_name = NULL)
}
\arguments{
\item{irace_results}{The data generated when loading the \code{Rdata} file created by \code{irace}}

\item{file_name}{String, file name to save plot (example: "~/path/to/file_name.png")}

\item{parameters}{String vector, the parameter names of parameter to be included in the plot
(example: param_names = c("algorithm","dlb"))}
}
\value{
sunburst plot
}
\description{
The isunburst function will return a sunburst plot of the selected parameters
}
\examples{
sampling_pie(iraceResults)
sampling_pie(iraceResults, param_names = c("algorithm", "dlb"))
}
