% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_by_iteration.R
\name{summarise_by_iteration}
\alias{summarise_by_iteration}
\title{Summarise by iteration}
\usage{
summarise_by_iteration(irace_results)
}
\arguments{
\item{irace_results}{The data generated when loading the \code{.Rdata} file created by \code{irace} (or the filename of that file).}
}
\value{
tibble
}
\description{
Summarise by iteration
}
\examples{
irace_result <- read_logfile(system.file(package="irace", "exdata",
                                         "irace-acotsp.Rdata", mustWork = TRUE))
summarise_by_iteration(irace_result)
}
