% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_test.R
\name{boxplot_test}
\alias{boxplot_test}
\title{Box Plot Testing}
\usage{
boxplot_test(irace_results, type = "all", rpd = TRUE, file_name = NULL)
}
\arguments{
\item{irace_results}{The data generated when loading the \code{Rdata} file created by \code{irace}}

\item{type}{String, either "all", "ibest" or "best". By default it is "all" which shows all the configurations,
"best" shows the configurations of the last iteration and
"ibest" shows the best configurations of each iteration}

\item{rpd}{Logical (default TRUE) TRUE to plot performance as the relative percentage deviation to 
best results, FALSE to plot raw performance}

\item{file_name}{String, File name to save plot (example: "~/patch/example/file_name.png")}
}
\value{
box plot
}
\description{
The function will return a box plot, using the data generated in the test
settings coloring the best configuration in each iteration
}
\examples{
boxplot_test(iraceResults)
boxplot_test(iraceResults, type = "ibest")
boxplot_test(iraceResults, type = "best")
}
