% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_training.R
\name{boxplot_training}
\alias{boxplot_training}
\title{Box Plot}
\usage{
boxplot_training(
  irace_results,
  number_iteration = NULL,
  id_configurations = NULL,
  rpd = TRUE,
  file_name = NULL
)
}
\arguments{
\item{irace_results}{The data generated when loading the \code{Rdata} file created by \code{irace}}

\item{number_iteration}{Numeric, iteration number that should be included in the plot (example: number_iteration = 5)}

\item{id_configurations}{Numeric vector, configurations ids whose performance should be included in the plot
(example: id_configurations = c(20,50,100,300,500,600,700))}

\item{rpd}{Logical (default TRUE) to plot performance as the relative percentage deviation to best results}

\item{file_name}{String, File name to save plot (example: "~/patch/example/file_name.png")}
}
\value{
box plot
}
\description{
Create a graphic of box plot of irace using the best quality id
By default the graph of the last iteration is shown
}
\examples{
boxplot_training(iraceResults)
boxplot_training(iraceResults, number_iteration = 5)
boxplot_training(iraceResults, id_configurations = c(20, 50, 100, 300, 500, 600, 700))
}
