% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configurations_display.R
\name{configurations_display}
\alias{configurations_display}
\title{The configurations by iteration and instance}
\usage{
configurations_display(
  irace_results,
  rpd = TRUE,
  file_name = NULL,
  .interactive = interactive()
)
}
\arguments{
\item{irace_results}{The data generated when loading the \code{Rdata} file created by \code{irace}}

\item{rpd}{logical(default TRUE) TRUE to plot performance as the relative percentage deviation to 
best results, FALSE to plot raw performance}

\item{file_name}{String, File name to save plot (example: `"~/patch/example/file_name.png"`)}

\item{.interactive}{Logical (Default `interactive()`), TRUE if the plot is generated interactively (plotly package) which
is the default option, or FALSE it is generated statically.}
}
\value{
plot
}
\description{
A graph is created with all the settings and instance of the training data
}
\examples{
\dontrun{ # Takes a long time.
configurations_display(iraceResults)
}
}
