% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configurations_display.R
\name{configurations_display}
\alias{configurations_display}
\title{The configurations by iteration and instance}
\usage{
configurations_display(
  irace_results,
  rpd = TRUE,
  filename = NULL,
  interactive = base::interactive()
)
}
\arguments{
\item{irace_results}{The data generated when loading the \code{Rdata} file created by \code{irace}}

\item{rpd}{(\code{logical(1)}) TRUE to plot performance as the relative
percentage deviation to best results per instance, FALSE to plot raw
performance.}

\item{filename}{(\code{character(1)}) File name to save the plot, for example \code{"~/patch/example/filename.png"}.}

\item{interactive}{(\code{logical(1)}) TRUE if the plot is generated
interactively (using \code{plotly}) or FALSE it is generated
statically. Defaults to the value returned by \code{interactive()},}
}
\value{
`ggplot()` object
}
\description{
A graph is created with all the settings and instance of the training data
}
\examples{
\dontrun{ # Takes a long time.
configurations_display(iraceResults)
}
}
