% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configurations_display.R
\name{configurations_display}
\alias{configurations_display}
\title{The configurations by iteration and instance}
\usage{
configurations_display(irace_results, rpd = TRUE, file_name = NULL)
}
\arguments{
\item{irace_results}{The data generated when loading the \code{Rdata} file created by \code{irace}}

\item{rpd}{Logical (default TRUE) to fit through an equation of minimum percentage distance
between the values of each row of all configurations}

\item{file_name}{String, A pdf will be created in the location and with the
assigned name (example: "~/patch/example/file_name")}
}
\value{
plot
}
\description{
A graph is created with all the settings and instance of the training data
}
\examples{
#configurations_display(iraceResults)
}
