% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_config.R
\name{distance_config}
\alias{distance_config}
\title{Distance between configurations}
\usage{
distance_config(iraceResults, idConfigurations, t = 0.05)
}
\arguments{
\item{iraceResults}{The data generated when loading the Rdata file created by irace}

\item{idConfigurations}{Numeric vector, you need to put the settings you
want to compare, only 2 values are allowed (example: idVector = c(806,809))}

\item{t}{Numeric, value to determine the range of difference between settings}
}
\value{
numeric
}
\description{
Distance between configurations
}
\examples{
NULL
}
