% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_config.R
\name{distance_config}
\alias{distance_config}
\title{Distance between configurations}
\usage{
distance_config(irace_results, id_configuration, t = 0.05)
}
\arguments{
\item{irace_results}{The data generated when loading the \code{.Rdata} file created by \code{irace} (or the filename of that file).}

\item{id_configuration}{Numeric, configuration id which should be compared to others
(example: id_configuration = c(806,809))}

\item{t}{Numeric, (default 0.05) threshold that defines the distance (percentage of the domain size)
to consider a parameter value equal to other.}
}
\value{
numeric
}
\description{
Calculate the difference between a configuration and the others in the irace data.
}
\examples{
NULL
}
