% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parameters_names.R
\name{get_parameters_names}
\alias{get_parameters_names}
\title{Parameters Names}
\usage{
get_parameters_names(irace_results)
}
\arguments{
\item{irace_results}{The data generated when loading the \code{Rdata} file created by \code{irace}}
}
\value{
String vector
}
\description{
The function returns a string vector with the names of
the document parameters irace_results
}
\examples{
get_parameters_names(iraceResults)
}
