% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iboxplot_test.R
\name{iboxplot_test}
\alias{iboxplot_test}
\title{Box Plot Testing}
\usage{
iboxplot_test(iraceResults, type = NULL, distance_min = FALSE, fileName = NULL)
}
\arguments{
\item{iraceResults}{The data generated when loading the Rdata file created by irace}

\item{type}{type = NULL, all testing experiments settings
type = 1, the last iteration of the elite settings
type = 2, the best settings of each iteration}

\item{distance_min}{The data is treated referring to the minimum distance of each percentage row}

\item{fileName}{It's of type string
You must place the path where you want to save the file and its name without the extension pdf}
}
\value{
plot
}
\description{
Box Plot Testing
}
\examples{
NULL
}
