% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibp.R
\name{ibp}
\alias{ibp}
\title{Graphic box plot}
\usage{
ibp(iraceResults, fileName = NULL, numberIteration = NULL)
}
\arguments{
\item{iraceResults}{The data generated when loading the Rdata file created by irace}

\item{fileName}{It is of type string.
You must place the path where you want to save the file and its name, it is not necessary to place the extension of the pdf}

\item{numberIteration}{It's of type numeric.
Allows to choose which iteration of the elite configuration of irace to graph}
}
\value{
plot
}
\description{
Create a graphic of box plot of irace using the best quality id
By default the graph of the last iteration is shown
}
\examples{
NULL
}
