% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibp.R
\name{ibp}
\alias{ibp}
\title{Box Plot}
\usage{
ibp(
  iraceResults,
  numberIteration = NULL,
  distance_min = FALSE,
  fileName = NULL
)
}
\arguments{
\item{iraceResults}{The data generated when loading the Rdata file created by irace}

\item{numberIteration}{Numeric, It is the number referring to the iteration that you want to graph,
the range of these varies according to the Rdata used (example: numberIteration = 5)}

\item{distance_min}{Logical (default FALSE) to fit through an equation of minimum percentage distance
between the values of each row of all configurations}

\item{fileName}{String, A pdf will be created in the location and with the
assigned name (example: "~/patch/example/filename")}
}
\value{
box plot
}
\description{
Create a graphic of box plot of irace using the best quality id
By default the graph of the last iteration is shown
}
\examples{
NULL
}
