% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iheatmap.R
\name{iheatmap}
\alias{iheatmap}
\title{Heat Map Plot}
\usage{
iheatmap(iraceResults, fileName = NULL)
}
\arguments{
\item{iraceResults}{The data generated when loading the Rdata file created by irace}

\item{fileName}{It's of type string
You must place the path where you want to save the file and its name without the extension pdf}
}
\value{
plot
}
\description{
Create a heat map plot interactive
}
\examples{
NULL
}
