% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iparallelcoord.R
\name{iparallelcoord}
\alias{iparallelcoord}
\title{Parallel Coordinates Plot static}
\usage{
iparallelcoord(
  iraceResults,
  idConfiguration = NULL,
  param_names = NULL,
  fileName = NULL
)
}
\arguments{
\item{iraceResults}{The data generated when loading the Rdata file created by irace}

\item{idConfiguration}{It is a vector with id values that you want to graph}

\item{param_names}{It's of type vector
is a vector with the parameters to be displayed in the plot}

\item{fileName}{It's of type string
You must place the path where you want to save the file and its name without the extension pdf}
}
\value{
plot
}
\description{
Parallel Coordinates Plot static
}
\examples{
NULL
}
