% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iparameter_density_iteration.R
\name{iparameter_density_iteration}
\alias{iparameter_density_iteration}
\title{Density plot based on its iteration}
\usage{
iparameter_density_iteration(iraceResults, parameter, fileName = NULL)
}
\arguments{
\item{iraceResults}{The data generated when loading the Rdata file created by irace}

\item{parameter}{It's of type string
The parameter you want to plot}

\item{fileName}{It's of type string
You must place the path where you want to save the file and its name without the extension pdf}
}
\value{
plot
}
\description{
Density plot based on its iteration
}
\examples{
NULL
}
