% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iparcoord.R
\name{iparcoord}
\alias{iparcoord}
\title{Parallel Coordinates Plot interactive}
\usage{
iparcoord(iraceResults, idConfiguration = NULL, param_names = NULL)
}
\arguments{
\item{iraceResults}{The data generated when loading the Rdata file created by irace}

\item{idConfiguration}{Numeric vector, you need to put the configurations you want to analyze
(example: idConfiguration = c(20,50,100,300,500,600,700))}

\item{param_names}{String vector, you need to put the parameters you want to analyze
(example: param_names = c("algorithm","alpha","rho","q0","rasrank"))}
}
\value{
parallel cordinates plot
}
\description{
The function will returns a parallel coordinates plot
interactive in shinyApp allowing the analysis of the set of parameters
allowing the visualization of the data and filter by iteration
}
\examples{
NULL
}
