% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iparcoord.R
\name{iparcoord}
\alias{iparcoord}
\title{Parallel Coordinates Plot}
\usage{
iparcoord(iraceResults, idIteration = NULL, param_names = NULL)
}
\arguments{
\item{iraceResults}{The data generated when loading the Rdata file created by irace}

\item{idIteration}{It is a vector with id values that you want to graph}

\item{param_names}{It's of type vector
is a vector with the parameters to be displayed in the plot}
}
\value{
plot
}
\description{
Parallel Coordinates Plot
}
\examples{
NULL
}
