% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iraceplot-package.R
\docType{package}
\name{iraceplot-package}
\alias{iraceplot}
\alias{iraceplot-package}
\title{iraceplot: Plots for Visualizing the Data Produced by the Irace Package}
\description{
Graphical visualization tools for analysing the data produced by irace.

Graphical Visualization Tools for Analysing the Data Produced by Irace.

boxplot_test;
scatter_test;
parallel_coord;
parallel_coord2;
parallel_cat;
sampling_frequency;
sampling_frequency_iteration;
sampling_pie;
plot_experiments_matrix;
sampling_distance;
sampling_heatmap;
sampling_heatmap2;
boxplot_training;
scatter_training;
boxplot_performance;
scatter_performance;
report;
plot_model;

If you need information about any function you can write:
?name_function

If you need more information, go to the following page:
https://auto-optimization.github.io/iraceplot/
}
\details{
License: MIT + file LICENSE
}
\seealso{
Useful links:
\itemize{
  \item \url{https://auto-optimization.github.io/iraceplot/}
  \item \url{https://github.com/auto-optimization/iraceplot/}
  \item Report bugs at \url{https://github.com/auto-optimization/iraceplot/issues}
}

}
\author{
\strong{Maintainer}: Leslie Pérez Cáceres \email{leslie.perez@pucv.cl} (\href{https://orcid.org/0000-0001-5553-6150}{ORCID})

Authors:
\itemize{
  \item Pablo Oñate Marín \email{pablo.onate.m@gmail.com}
  \item Manuel López-Ibáñez \email{manuel.lopez-ibanez@uma.es} (\href{https://orcid.org/0000-0001-9974-1295}{ORCID})
}


Maintainers: Pablo Oñate Marín and Leslie Pérez Cáceres and Manuel López-Ibañez
        \email{leslie.perez@pucv.cl}
}
\keyword{automatic}
\keyword{configuration}
\keyword{internal}
\keyword{package}
\keyword{plot}
