% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iraceplot-package.R
\docType{package}
\name{iraceplot-package}
\alias{iraceplot}
\alias{iraceplot-package}
\title{iraceplot: Plots for visualizing the data produced by the irace package}
\description{
Graphical visualization tools for analysing the data produced by irace.

Graphical visualization tools for analysing the data produced by irace.

boxplot_test;
scatter_test;
parallel_coord;
parallel_coord2;
parallel_cat;
sampling_frequency;
sampling_frequency_iteration;
sampling_pie;
plot_experiments_matrix;
sampling_distance;
boxplot_training;
scatter_training;
report;
plot_model;

If you need information about any function you can write:
?name_function

If you need more information, go to the following page:
https://pabloonate.github.io/iraceplot/
}
\details{
License: MIT + file LICENSE
}
\seealso{
Useful links:
\itemize{
  \item \url{https://auto-optimization.github.io/iraceplot}
  \item \url{https://github.com/auto-optimization/iraceplot}
  \item Report bugs at \url{https://github.com/auto-optimization/iraceplot/issues}
}

}
\author{
\strong{Maintainer}: Leslie Pérez Cáceres \email{leslie.perez@pucv.cl}

Authors:
\itemize{
  \item Pablo Oñate Marín \email{pablo.onate.m@gmail.com}
  \item Manuel López-Ibáñez \email{manuel.lopez-ibanez@uma.es} (\href{https://orcid.org/0000-0001-9974-1295}{ORCID})
}


Maintainers: Pablo Oñate Marín and Leslie Pérez Cáceres
        \email{pablo.onate.m@gmail.com}
}
\keyword{automatic}
\keyword{configuration}
\keyword{internal}
\keyword{package}
\keyword{plot}
