% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iraceplot-package.R
\docType{package}
\name{iraceplot-package}
\alias{iraceplot}
\alias{iraceplot-package}
\title{iraceplot: Plots for visualizing the data produced by the irace package}
\description{
For the creation of various types of graphs that facilitates to understand the data produced by irace.

The different functions used in iraceplot will allow the creation of different graphics. The functions are as follows:

boxplot_test;
scatter_test;
parallel_coord;
parallel_coord2;
parallel_cat;
sampling_frequency;
sampling_frequency_iteration;
sampling_pie;
plot_experiments_matrix;
sampling_distance;
boxplot_training;
scatter_training;
report;
plot_model;

If you need information about any function you can put:
?name_function

If you need more information, go to the following page:
https://pabloonate.github.io/iraceplot/
}
\details{
License: MIT + file LICENSE
}
\seealso{
Useful links:
\itemize{
  \item \url{https://auto-optimization.github.io/iraceplot/}
}

}
\author{
\strong{Maintainer}: Pablo Oñate Marín \email{pablo.onate.m@gmail.com}

Authors:
\itemize{
  \item Leslie Pérez Cáceres \email{leslie.perez@pucv.cl}
}


Maintainers: Pablo Oñate Marín and Leslie Pérez Cáceres
        \email{pablo.onate.m@gmail.com}
}
\keyword{automatic}
\keyword{configuration}
\keyword{internal}
\keyword{package}
\keyword{plot}
