% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iraceplot-package.R
\docType{package}
\name{iraceplot-package}
\alias{iraceplot}
\alias{iraceplot-package}
\title{iraceplot: Plots for Visualizing the Data Produced by the 'irace' Package}
\description{
Graphical visualization tools for analyzing the data produced by 'irace'. The 'iraceplot' package enables users to analyze the performance and the parameter space data sampled by the configuration during the search process. It provides a set of functions that generate different plots to visualize the configurations sampled during the execution of 'irace' and their performance. The functions just require the log file generated by 'irace' and, in some cases, they can be used with user-provided data.

Graphical Visualization Tools for Analysing the Data Produced by irace.

boxplot_performance;
boxplot_test;
boxplot_training;
parallel_cat;
parallel_coord2;
parallel_coord;
plot_experiments_matrix;
plot_model;
report;
sampling_distance;
sampling_frequency;
sampling_frequency_iteration;
sampling_heatmap2;
sampling_heatmap;
sampling_pie;
scatter_performance;
scatter_test;
scatter_training;

If you need information about any function you can write:
?name_function

If you need more information, go to the following page:
https://auto-optimization.github.io/iraceplot/
}
\details{
License: MIT + file LICENSE
}
\seealso{
Useful links:
\itemize{
  \item \url{https://auto-optimization.github.io/iraceplot/}
  \item \url{https://github.com/auto-optimization/iraceplot/}
  \item Report bugs at \url{https://github.com/auto-optimization/iraceplot/issues}
}

}
\author{
\strong{Maintainer}: Manuel López-Ibáñez \email{manuel.lopez-ibanez@manchester.ac.uk} (\href{https://orcid.org/0000-0001-9974-1295}{ORCID})

Authors:
\itemize{
  \item Pablo Oñate Marín \email{pablo.onate.m@gmail.com}
  \item Leslie Pérez Cáceres \email{leslie.perez@pucv.cl} (\href{https://orcid.org/0000-0001-5553-6150}{ORCID})
}


Maintainers: Pablo Oñate Marín and Leslie Pérez Cáceres and Manuel López-Ibañez
\email{leslie.perez@pucv.cl}
}
\keyword{automatic}
\keyword{configuration}
\keyword{internal}
\keyword{package}
\keyword{plot}
