% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iraceplot-package.R
\docType{package}
\name{iraceplot-package}
\alias{iraceplot}
\alias{iraceplot-package}
\title{The iraceplot package: \packageTitle{iraceplot}}
\description{
\packageDescription{iraceplot}
}
\details{
boxplot_performance;
boxplot_test;
boxplot_training;
parallel_cat;
plot_configurations;
parallel_coord;
plot_experiments_matrix;
plot_model;
report;
sampling_distance;
sampling_frequency;
sampling_frequency_iteration;
sampling_heatmap2;
sampling_heatmap;
sampling_pie;
scatter_performance;
scatter_test;
scatter_training;

If you need information about any function you can write: \code{?name_function}

If you need more information, go to the following page:
https://auto-optimization.github.io/iraceplot/
}
\seealso{
Useful links:
\itemize{
  \item \url{https://auto-optimization.github.io/iraceplot/}
  \item \url{https://github.com/auto-optimization/iraceplot/}
  \item Report bugs at \url{https://github.com/auto-optimization/iraceplot/issues}
}

}
\author{
\strong{Maintainer}: Manuel López-Ibáñez \email{manuel.lopez-ibanez@manchester.ac.uk} (\href{https://orcid.org/0000-0001-9974-1295}{ORCID})

Authors:
\itemize{
  \item Pablo Oñate Marín \email{pablo.onate.m@gmail.com}
  \item Leslie Pérez Cáceres \email{leslie.perez@pucv.cl} (\href{https://orcid.org/0000-0001-5553-6150}{ORCID})
}

}
\keyword{automatic}
\keyword{configuration}
\keyword{package}
\keyword{plot}
