% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iscatter.R
\name{iscatter}
\alias{iscatter}
\title{Scatter Plot Training}
\usage{
iscatter(iraceResults, idVector, rpd = TRUE, fileName = NULL)
}
\arguments{
\item{iraceResults}{The data generated when loading the Rdata file created by irace}

\item{idVector}{Numeric vector, you need to put the elite settings you
want to compare, only 2 values are allowed (example: idVector = c(806,809))}

\item{rpd}{Logical (default TRUE) to fit through an equation of minimum percentage
distance between the values of each row of all configurations}

\item{fileName}{String, A pdf will be created in the location and with the
assigned name (example: "~/patch/example/filename")}
}
\value{
plot
}
\description{
The function will return a scatter plot
comparing two elite configurations in training
}
\examples{
NULL
}
