% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iscatter.R
\name{iscatter}
\alias{iscatter}
\title{Graphic Scatter}
\usage{
iscatter(iraceResults, idVector, fileName = NULL)
}
\arguments{
\item{iraceResults}{The data generated when loading the Rdata file created by irace}

\item{idVector}{It is a vector with 2 id values that you want to graph}

\item{fileName}{It's of type string
You must place the path where you want to save the file and its name without the extension pdf}
}
\value{
plot
}
\description{
Create a graphic of scatter of irace using a vector with 2 values referring to the id
}
\examples{
NULL
}
