% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_cat.R
\name{parallel_cat}
\alias{parallel_cat}
\title{Parallel Cordinate Category}
\usage{
parallel_cat(
  irace_results,
  id_configuration = NULL,
  param_names = NULL,
  iterations = NULL,
  plot_all_parameters = FALSE,
  file_name = NULL
)
}
\arguments{
\item{irace_results}{The data generated when loading the \code{Rdata} file created by \code{irace}}

\item{id_configuration}{Numeric vector, configuration ids to be included in the plot
(example: id_configuration = c(20,50,100,300,500,600,700))}

\item{param_names}{String vector, parameters to be included in the plot
(example: param_names = c("algorithm","alpha","rho","q0","rasrank"))}

\item{iterations}{NUmeric vector, iterations from which configuration should be obtained
(example: iterations = c(1,4,5))}

\item{plot_all_parameters}{logical (default FALSE), TRUE if all parameter should be included or 
FALSE if the first set of 15 parameters provided or in the data file should
be displayed.}

\item{file_name}{String, file name to save plot (example: "~/path/to/file_name.png")}
}
\value{
parallel categories plot
}
\description{
The parallel_cat function will return a graph of categorical
parallel coordinates allowing selecting configurations by id or iteration
and selecting the parameter to be included in the plot
}
\examples{
parallel_cat(iraceResults)
parallel_cat(iraceResults, id_configuration = c(20, 50, 100, 300, 500, 600, 700))
parallel_cat(iraceResults, param_names = c("algorithm", "alpha", "rho", "q0", "rasrank"))
parallel_cat(iraceResults, iterations = c(1, 4, 6))
}
