% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_coord.R
\name{parallel_coord}
\alias{parallel_coord}
\title{Parallel Coordinates Plot static}
\usage{
parallel_coord(
  iraceResults,
  idConfiguration = NULL,
  param_names = NULL,
  iterations = NULL,
  fileName = NULL
)
}
\arguments{
\item{iraceResults}{The data generated when loading the Rdata file created by irace}

\item{idConfiguration}{Numeric vector, you need to put the configurations you want to analyze
(example: idConfiguration = c(20,50,100,300,500,600,700))}

\item{param_names}{String vector, you need to put the parameters you want to analyze
(example: param_names = c("algorithm","alpha","rho","q0","rasrank"))}

\item{iterations}{NUmeric vector, you need to put the iterations you want to analyze
(example: iterations = c(1,4,5))}

\item{fileName}{A pdf will be created in the location and with
the assigned name (example: "~/patch/example/filename")}
}
\value{
parallel cordinates plot
}
\description{
The iparallelcoord function will return a parallel cordinates plot
allowing the analysis of the set of parameters
}
\examples{
NULL
}
