% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_experiments_matrix.R
\name{plot_experiments_matrix}
\alias{plot_experiments_matrix}
\title{Heat Map Plot}
\usage{
plot_experiments_matrix(
  irace_results,
  file_name = NULL,
  .interactive = interactive()
)
}
\arguments{
\item{irace_results}{The data generated when loading the \code{Rdata} file created by \code{irace}}

\item{file_name}{string, A pdf will be created in the location and with the
assigned name (example: "~/patch/example/file_name")}

\item{.interactive}{Logical (Default interactive()), TRUE if the plot is generated interactively (plotly package) which
is the default option, or FALSE it is generated statically.}
}
\value{
heatmap plot
}
\description{
The function will return a heat map plot of all experimental data configurations
}
\examples{
plot_experiments_matrix(iraceResults, .interactive = interactive())
}
