% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_parameters.R
\name{printParameters}
\alias{printParameters}
\title{Print parameter space in the textual format accepted by irace.}
\usage{
printParameters(params)
}
\arguments{
\item{params}{(\code{list()}) Parameter object stored in \code{irace.Rdata} or read with \code{irace::readParameters()}.}
}
\description{
Print parameter space in the textual format accepted by irace.
}
\examples{
load(system.file(package="iraceplot", "exdata", "guide-example.Rdata", mustWork = TRUE))
printParameters(iraceResults$parameters)
}
