% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{Report Iraceplot}
\usage{
report(irace_results, file_name)
}
\arguments{
\item{irace_results}{The data generated when loading the \code{Rdata} file created by \code{irace}}

\item{file_name}{String, file name indicating where to save the report (example: "~/path-to/file_name")}
}
\value{
document
}
\description{
The `report` function creates an HTML report of the most relevant irace log
data.
}
\examples{
NULL
}
