% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{Report Iraceplot}
\usage{
report(iraceResults, fileName = NULL)
}
\arguments{
\item{iraceResults}{irace log variable obtained from the Rdata file generate by irace}

\item{fileName}{String, file name indicating where to save a pdf file with the plot.
A pdf extension will be added to the file name provided
(example: "~/path-to/filename")}
}
\value{
document
}
\description{
Report Iraceplot
}
\examples{
NULL
}
