% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{Report Iraceplot}
\usage{
report(format = "html", irace_results = NULL, irace_file = NULL, file_name)
}
\arguments{
\item{format}{String, either "pdf" or "html". The parameter according to its
value is how the output file format will be. By default it is "html"}

\item{irace_results}{The data generated when loading the \code{Rdata} file created by \code{irace}}

\item{irace_file}{String, indicate the location where the Rdata file is located
(example: "~/path-to/file.Rdata")}

\item{file_name}{String, file name indicating where to save a pdf file with the plot.
A pdf extension will be added to the file name provided
(example: "~/path-to/file_name")}
}
\value{
document
}
\description{
Report Iraceplot
}
\examples{
NULL
}
