% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{Create HTML Report from irace data}
\usage{
report(irace_results, filename = "report", interactive = base::interactive())
}
\arguments{
\item{irace_results}{The data generated when loading the \code{Rdata} file created by \code{irace}}

\item{filename}{(\code{character(1)})
Filename indicating where to save the report (example: \code{"~/path-to/filename"})}

\item{interactive}{(\code{logical(1)}) TRUE if the plot is generated
interactively (using \code{plotly}) or FALSE it is generated
statically. Defaults to the value returned by \code{interactive()},}
}
\description{
This function creates an HTML report of the most relevant irace data.  This
report provides general statistics and plots that show the best
configurations and their performance. Example: \url{https://auto-optimization.github.io/iraceplot/articles/example/report_example.html}
}
\examples{
# load(system.file(package="irace", "exdata", "irace-acotsp.Rdata", mustWork = TRUE))
# report(iraceResults, filename="report")
NULL
}
