% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_density.R
\name{sampling_density}
\alias{sampling_density}
\title{Density plot based on its iteration}
\usage{
sampling_density(irace_results, parameter, file_name = NULL)
}
\arguments{
\item{irace_results}{The data generated when loading the \code{Rdata} file created by \code{irace}}

\item{parameter}{String, value of the parameter to be analyzed (example: parameter = "alpha")}

\item{file_name}{String, A pdf will be created in the location and with the
assigned name (example: "~/patch/example/file_name")}
}
\value{
density plot
}
\description{
The function will return a density plot for parameters numeric data
in case of placing a categorical parameter, it will show a bar plot
}
\examples{
sampling_density(iraceResults, parameter = "alpha")
}
