% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_pie.R
\name{sampling_pie}
\alias{sampling_pie}
\title{Sunburst Plot}
\usage{
sampling_pie(irace_results, parameters = NULL, file_name = NULL)
}
\arguments{
\item{irace_results}{The data generated when loading the Rdata file created by irace}

\item{parameters}{String vector, a set of categorical type parameters
(example: parameters = c("algorithm","dlb"))}

\item{file_name}{String, A pdf will be created in the location and with the assigned
name (example: "~/patch/example/file_name")}
}
\value{
sunburst plot
}
\description{
The isunburst function will return a sunburst plot of the categorical parameters
}
\examples{
sampling_pie(iraceResults)
sampling_pie(iraceResults, parameters = c("algorithm","dlb"))
}
