% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_pie.R
\name{sampling_pie}
\alias{sampling_pie}
\title{Sunburst Plot}
\usage{
sampling_pie(iraceResults, dependency = FALSE, fileName = NULL)
}
\arguments{
\item{iraceResults}{The data generated when loading the Rdata file created by irace}

\item{dependency}{Logical (default FALSE) that allows to verify if the parameters
are dependent on others, modifying the visualization of the plot}

\item{fileName}{String, A pdf will be created in the location and with the assigned
name (example: "~/patch/example/filename")}
}
\value{
sunburst plot
}
\description{
The isunburst function will return a sunburst plot of the categorical parameters
}
\examples{
NULL
}
