% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_test.R
\name{scatter_test}
\alias{scatter_test}
\title{Scatter Plot Testing}
\usage{
scatter_test(
  irace_results,
  id_configurations,
  rpd = TRUE,
  filename = NULL,
  interactive = base::interactive()
)
}
\arguments{
\item{irace_results}{The data generated when loading the \code{Rdata} file created by \code{irace}}

\item{id_configurations}{String vector, configuration ids whose performance should be displayed
(example: id_configurations = c("92","119"))}

\item{rpd}{(\code{logical(1)}) TRUE to plot performance as the relative
percentage deviation to best results per instance, FALSE to plot raw
performance.}

\item{filename}{(\code{character(1)}) File name to save the plot, for example \code{"~/path/example/filename.png"}.}

\item{interactive}{(\code{logical(1)}) TRUE if the plot is generated
interactively (using \code{plotly}) or FALSE it is generated
statically. Defaults to the value returned by \code{interactive()},}
}
\value{
\code{ggplot()} object
}
\description{
The \code{scatter_test} function creates a scatter plot comparing the performance of two
configurations in the test instances. Each point in the plot represents an instance
and the color of the points indicates if one configuration is better than the other.

The performance data is obtained from the test evaluations performed by irace.
Note that the testing is not a default feature in irace and should be enabled in
the setup.

Configuration ids provided in \code{id_configurations} should belong to
elite configuration set evaluated in the test (see the irace package user
guide for more details).
}
\examples{
scatter_test(iraceResults, id_configurations = c("92", "119"))
\dontrun{ 
scatter_test(iraceResults, id_configurations = c("92", "119"), rpd=FALSE)
}
}
