% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_test.R
\name{scatter_test}
\alias{scatter_test}
\title{Scatter Plot Testing}
\usage{
scatter_test(irace_results, id_configurations, rpd = TRUE, file_name = NULL)
}
\arguments{
\item{irace_results}{The data generated when loading the \code{Rdata} file created by \code{irace}}

\item{id_configurations}{String vector, you need to put the elite settings you
want to compare, only 2 values are allowed (example: idVector = c("92","119"))}

\item{rpd}{logical(default TRUE) to fit through an equation of minimum percentage distance
between the values of each row of all configurations}

\item{file_name}{String, A pdf will be created in the location and with the assigned
name (example: "~/patch/example/file_name")}
}
\value{
scatter plot
}
\description{
The function will return a scatter plot
comparing two elite configurations in testing
}
\examples{
scatter_test(iraceResults, id_configurations = c("92","119"))
}
