% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_training.R
\name{scatter_training}
\alias{scatter_training}
\title{Scatter Plot Training}
\usage{
scatter_training(
  irace_results,
  id_configurations,
  rpd = TRUE,
  file_name = NULL,
  .interactive = interactive()
)
}
\arguments{
\item{irace_results}{The data generated when loading the \code{Rdata} file created by \code{irace}}

\item{id_configurations}{Numeric vector,  configuration ids whose performance should be displayed (example: id_configurations = c("92","119"))}

\item{rpd}{logical(default TRUE) TRUE to plot performance as the relative percentage deviation to 
best results, FALSE to plot raw performance}

\item{file_name}{String, File name to save plot (example: "~/patch/example/file_name.png")}

\item{.interactive}{Logical (Default interactive()), TRUE if the plot is generated interactively (plotly package) which
is the default option, or FALSE it is generated statically.}
}
\value{
scatter plot
}
\description{
The function will return a scatter plot
comparing two configurations in the training performance data
}
\examples{
scatter_training(iraceResults, id_configurations = c(806, 809))
}
