% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_training.R
\name{scatter_training}
\alias{scatter_training}
\title{Scatter Plot Training}
\usage{
scatter_training(
  irace_results,
  id_configurations,
  rpd = TRUE,
  filename = NULL,
  interactive = base::interactive()
)
}
\arguments{
\item{irace_results}{The data generated when loading the \code{Rdata} file created by \code{irace}}

\item{id_configurations}{Configuration ids to be included in the
plot. Example: \code{c(20,50,100,300,500,600,700)}}

\item{rpd}{(\code{logical(1)}) TRUE to plot performance as the relative
percentage deviation to best results per instance, FALSE to plot raw
performance.}

\item{filename}{(\code{character(1)}) File name to save the plot, for example \code{"~/path/example/filename.png"}.}

\item{interactive}{(\code{logical(1)}) TRUE if the plot is generated
interactively (using \code{plotly}) or FALSE it is generated
statically. Defaults to the value returned by \code{interactive()},}
}
\value{
\code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Create a scatter plot that displays the performance of two configurations on
the training performance. Each point in the plot represents an instance and
the color of the points indicates if one configuration is better than the
other.
}
\details{
The performance data is obtained from the evaluations performed by irace
during the execution process, consequently the number of evaluations
can differ between configurations due to the elimination process applied by
irace. This plot only shows performance data only for instances in which both
configurations are executed.
}
\examples{
scatter_training(iraceResults, id_configurations = c(806, 809))
\dontrun{ 
scatter_training(iraceResults, id_configurations = c(806, 809), rpd = FALSE)
}
}
